refactor: components/を画面ベースのディレクトリ構成に再編成
- screens/, leftPane/, rightPane/ を廃止
- import/: ImportScreen, Dropzone, GettingStarted, SavedFiles
- aggregation/: AggregationScreen, ResultView, CrossConfig, 各テーブル/チャート
- header/: SettingsModal
- shared/: StatusDot, ToggleButton（複数箇所から使用）
- 未使用のescHtml.tsを削除
