+++ src/components/aggregation/AggregationScreen.tsx
@@ -1,11 +1,11 @@
 import { useState } from "preact/hooks";
-import CrossConfig from "../../leftPane/CrossConfig";
-import ResultView from "../../rightPane/ResultView";
-import { runDuckDBAggregation } from "../../../lib/duckdbBridge";
-import { buildQuestionDefs, type LayoutMeta } from "../../../lib/layout";
-import { questionKey, type AggResult, type QuestionDef } from "../../../lib/agg/aggregate";
-import { t } from "../../../lib/i18n";
-import { ToggleButton, ToggleGroup } from "../../shared/ToggleButton";
+import CrossConfig from "./CrossConfig";
+import ResultView from "./ResultView";
+import { runDuckDBAggregation } from "../../lib/duckdbBridge";
+import { buildQuestionDefs, type LayoutMeta } from "../../lib/layout";
+import { questionKey, type AggResult, type QuestionDef } from "../../lib/agg/aggregate";
+import { t } from "../../lib/i18n";
+import { ToggleButton, ToggleGroup } from "../shared/ToggleButton";
 
 type CsvData = { text: string; fileName: string; headers: string[]; rowCount: number };
 type LayoutData = { json: string; fileName: string; meta: LayoutMeta };
