+++ src/components/aggregation/ChartRenderer.tsx
