+++ src/components/aggregation/GtTable.tsx
