+++ src/components/aggregation/ResultView.tsx
