--- src/components/shared/escHtml.ts
@@ -1,3 +1,0 @@
-export function escHtml(str: string): string {
-  return String(str).replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;");
-}
